Refactor level and member routes; update sanitization middleware; modify course and level entities; add Ormo config to .gitignore; remove unused users HTTP file

+++ src/shared/orm.ts
@@ -18,9 +18,10 @@ export const orm = await MikroORM.init({
 });
 export const syncSchema = async () => {
   const generator = orm.getSchemaGenerator();
-  /*
+  let y = false ;
+  if(y){
         await generator.dropSchema();
         await generator.createSchema();
-        */
+  }
   await generator.updateSchema();
 };
